Split adding derived values and ngs data into separate functions
