+++ sources2csr/sources2csr.py
@@ -7,7 +7,7 @@ import click
 from csr2tsv.subject_registry_writer import SubjectRegistryWriter
 
 from csr2tsv.study_registry_writer import StudyRegistryWriter
-from sources2csr.derived_values import add_derived_values
+from sources2csr.helper_variables import add_derived_values, add_helper_variables
 from sources2csr.legacy.legacy_sources_reader import LegacySourcesReader
 from sources2csr.sources_reader import SourcesReader
 
@@ -42,7 +42,7 @@ def sources2csr(input_dir, output_dir, config_dir):
         reader = SourcesReader(input_dir=input_dir, output_dir=output_dir, config_dir=config_dir)
         subject_registry = reader.read_subject_data()
         ngs_set = reader.read_ngs_data()
-        add_derived_values(subject_registry, ngs_set)
+        add_helper_variables(subject_registry, ngs_set)
         subject_registry_writer = SubjectRegistryWriter(output_dir)
         subject_registry_writer.write(subject_registry)
